almost all fixes
