+++ CountriesListApp/CountriesList/View/CountriesListViewController.swift
@@ -16,8 +16,8 @@ final class CountriesListViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.presenter?.view = self
-        self.presenter?.getData()
+        presenter?.view = self
+        presenter?.getData()
     }
 
     private func printTest() {
@@ -26,18 +26,18 @@ final class CountriesListViewController: UIViewController {
 
     private func showAlert(_ error: String) {
         let alert = UIAlertController(title: "Error", message: error, preferredStyle: .alert)
-        self.present(alert, animated: true)
+        present(alert, animated: true)
     }
 }
 
 extension CountriesListViewController: CountriesListProtocol {
 
     func success(data: [Country]) {
-        self.countries = data
-        self.printTest()
+        countries = data
+        printTest()
     }
 
     func failure(error: Error) {
-        self.showAlert(error.localizedDescription)
+        showAlert(error.localizedDescription)
     }
 }
