+++ CountriesListApp/Helpers/CountriesProtocols.swift
@@ -1,0 +1,28 @@
+//
+//  CountriesProtocols.swift
+//  CountriesListApp
+//
+//  Created by Григорий Душин on 22.12.2023.
+//
+
+import Foundation
+
+protocol CountriesListProtocol: AnyObject {
+    func success(data: [Country])
+    func failure(error: Error)
+}
+
+protocol CountriesListPresenterProtocol: AnyObject {
+    func getData()
+    var view: CountriesListProtocol? { get set }
+}
+
+protocol CountryLoaderProtocol {
+    func countryDataLoad(completion: @escaping (Result<[Country], Error>) -> Void)
+}
+
+protocol DataLoadable {
+    associatedtype DataType: Decodable
+    func loadData(completion: @escaping (Result<[DataType], Error>) -> Void)
+}
+
